+++ src/stagecraft/core/types.py
@@ -2,12 +2,9 @@ from typing import TypeVar
 
 import numpy as np
 from numpy.typing import NDArray
-from pandera.typing import DataFrame
 from typing_extensions import TypeAlias
 
 T = TypeVar("T")
-
-PaDataFrame: TypeAlias = DataFrame[T]
 
 NDArrayGen: TypeAlias = NDArray[np.generic]
 NDArrayStr: TypeAlias = NDArray[np.str_]
@@ -25,7 +22,6 @@ NDArrayFloat: TypeAlias = NDArrayFloat64
 
 
 __all__ = [
-    "PaDataFrame",
     "NDArrayGen",
     "NDArrayStr",
     "NDArrayBool",